move diagram jobs and associated listener to new (non API) package
